Remove more uses of isizeof macro
These are WIN32 conditional code so got missed before.

+++ compiler/Generic/GrCommon.h
@@ -73,13 +73,7 @@ typedef schar	achar;
 typedef achar 	*Psz;
 //typedef const achar * Pcsz;
 
-
-
-// This is to make a signed isizeof operator, otherwise we get tons of warnings about
-// signed / unsigned mismatches.
-
 #define SizeOfArray(rgt) (sizeof(rgt) / sizeof *rgt)
-
 
 /***********************************************************************************************
 	Tests for valid strings and pointers.
@@ -112,12 +106,12 @@ inline bool ValidReadPtrSize(const void *pv, size_t cb)
 	return pv != 0 && !GrIsBadReadPtr(pv, cb);
 }
 
-inline bool ValidWritePtrSize(void *pv, int cb)
+inline bool ValidWritePtrSize(void *pv, size_t cb)
 {
-//	if (!bstr || ::IsBadReadPtr((byte *)bstr - isizeof(int), isizeof(int) + isizeof(OLECHAR)))
+//	if (!bstr || ::IsBadReadPtr((byte *)bstr - sizeof(int), sizeof(int) + sizeof(OLECHAR)))
 //		return false;
 //	int cb = ((int *)bstr)[-1];
-//	if (::IsBadReadPtr((byte *)bstr - isizeof(int), isizeof(int) + isizeof(OLECHAR) + cb))
+//	if (::IsBadReadPtr((byte *)bstr - sizeof(int), sizeof(int) + sizeof(OLECHAR) + cb))
 //		return false;
 	if (cb < 0)	return false;
 	if (cb == 0)	return true;
